requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
distro==1.9.0
Django==4.2.23
django-agora==0.4
django-cloudinary-storage==0.3.0
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django_notifications==0.1.dev0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonfield==3.2.0
openai==1.101.0
packaging==24.2
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-decouple==3.8
pytz==2026.1.post1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
swapper==1.4.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.12.0
